Added job progress; refactored jobmonitor to be strategy pattern; changed the handlers to stateless

+++ src/Common/DTO/Job.cs
@@ -38,7 +38,7 @@ namespace Microsoft.HpcAcm.Common.Dto
         public DateTimeOffset LastChangedAtAt { get; set; }
         public JobState State { get; set; }
         public JobType Type { get; set; }
-        public int Progress { get; set; }
+        public double Progress { get; set; }
         public int RequeueCount { get; set; } = 0;
         public bool FailJobOnTaskFailure { get; set; } = false;
